+++ tests/testthat/test-scan_packages.R
@@ -59,10 +59,10 @@ test_that("returns all package dependencies when dependencies = TRUE", {
   pkgs.df <- scan_packages(pkgs = "grateful", dependencies = TRUE)
 
   expect_identical(pkgs.df$pkg,
-                   c("base64enc", "bslib", "cachem", "digest", "evaluate",
+                   c("R6", "base64enc", "bslib", "cachem", "digest", "evaluate",
                      "fastmap", "fontawesome", "fs", "glue", "grateful",
                      "highr", "htmltools", "jquerylib", "knitr", "lifecycle",
-                     "memoise", "mime", "R6", "rappdirs", "remotes", "renv",
+                     "memoise", "mime", "rappdirs", "remotes", "renv",
                      "rmarkdown", "sass", "tidyverse", "tinytex", "xfun", "yaml")
                    )